# standard/generatedCode/wfm/sbi/server.gen.go
// Package sbi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package sbi

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Retrieve bundle information for a specific device and digest
	// (GET /api/v1/clients/{clientId}/bundles/{digest})
	GetApiV1ClientsClientIdBundlesDigest(ctx echo.Context, clientId string, digest string, params GetApiV1ClientsClientIdBundlesDigestParams) error
	// Report device capabilities
	// (POST /api/v1/clients/{clientId}/capabilities)
	PostApiV1ClientsClientIdCapabilities(ctx echo.Context, clientId string) error
	// Update device capabilities (Update)
	// (PUT /api/v1/clients/{clientId}/capabilities)
	PutApiV1ClientsClientIdCapabilities(ctx echo.Context, clientId string) error
	// Retrieve the complete desired state for all workloads assigned to a device
	// (GET /api/v1/clients/{clientId}/deployments)
	GetApiV1ClientsClientIdDeployments(ctx echo.Context, clientId string, params GetApiV1ClientsClientIdDeploymentsParams) error
	// Report deployment status
	// (POST /api/v1/clients/{clientId}/deployments/{deploymentId}/status)
	PostApiV1ClientsClientIdDeploymentsDeploymentIdStatus(ctx echo.Context, clientId string, deploymentId string) error
	// Retrieve an individual ApplicationDeployment YAML file
	// (GET /api/v1/clients/{clientId}/deployments/{deploymentId}/{digest})
	GetApiV1ClientsClientIdDeploymentsDeploymentIdDigest(ctx echo.Context, clientId string, deploymentId string, digest string, params GetApiV1ClientsClientIdDeploymentsDeploymentIdDigestParams) error
	// Complete onboarding with client certificate
	// (POST /api/v1/onboarding)
	PostApiV1Onboarding(ctx echo.Context) error
	// Download Root CA certificate
	// (GET /api/v1/onboarding/certificate)
	GetApiV1OnboardingCertificate(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetApiV1ClientsClientIdBundlesDigest converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiV1ClientsClientIdBundlesDigest(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "clientId" -------------
	var clientId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "clientId", runtime.ParamLocationPath, ctx.Param("clientId"), &clientId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter clientId: %s", err))
	}

	// ------------- Path parameter "digest" -------------
	var digest string

	err = runtime.BindStyledParameterWithLocation("simple", false, "digest", runtime.ParamLocationPath, ctx.Param("digest"), &digest)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter digest: %s", err))
	}

	ctx.Set(PayloadSignatureScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiV1ClientsClientIdBundlesDigestParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "If-None-Match" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("If-None-Match")]; found {
		var IfNoneMatch string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for If-None-Match, got %d", n))
		}

		err = runtime.BindStyledParameterWithLocation("simple", false, "If-None-Match", runtime.ParamLocationHeader, valueList[0], &IfNoneMatch)
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter If-None-Match: %s", err))
		}

		params.IfNoneMatch = &IfNoneMatch
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiV1ClientsClientIdBundlesDigest(ctx, clientId, digest, params)
	return err
}

// PostApiV1ClientsClientIdCapabilities converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiV1ClientsClientIdCapabilities(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "clientId" -------------
	var clientId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "clientId", runtime.ParamLocationPath, ctx.Param("clientId"), &clientId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter clientId: %s", err))
	}

	ctx.Set(PayloadSignatureScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiV1ClientsClientIdCapabilities(ctx, clientId)
	return err
}

// PutApiV1ClientsClientIdCapabilities converts echo context to params.
func (w *ServerInterfaceWrapper) PutApiV1ClientsClientIdCapabilities(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "clientId" -------------
	var clientId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "clientId", runtime.ParamLocationPath, ctx.Param("clientId"), &clientId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter clientId: %s", err))
	}

	ctx.Set(PayloadSignatureScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutApiV1ClientsClientIdCapabilities(ctx, clientId)
	return err
}

// GetApiV1ClientsClientIdDeployments converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiV1ClientsClientIdDeployments(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "clientId" -------------
	var clientId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "clientId", runtime.ParamLocationPath, ctx.Param("clientId"), &clientId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter clientId: %s", err))
	}

	ctx.Set(PayloadSignatureScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiV1ClientsClientIdDeploymentsParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "If-None-Match" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("If-None-Match")]; found {
		var IfNoneMatch string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for If-None-Match, got %d", n))
		}

		err = runtime.BindStyledParameterWithLocation("simple", false, "If-None-Match", runtime.ParamLocationHeader, valueList[0], &IfNoneMatch)
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter If-None-Match: %s", err))
		}

		params.IfNoneMatch = &IfNoneMatch
	}
	// ------------- Optional header parameter "Accept" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Accept")]; found {
		var Accept string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for Accept, got %d", n))
		}

		err = runtime.BindStyledParameterWithLocation("simple", false, "Accept", runtime.ParamLocationHeader, valueList[0], &Accept)
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter Accept: %s", err))
		}

		params.Accept = &Accept
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiV1ClientsClientIdDeployments(ctx, clientId, params)
	return err
}

// PostApiV1ClientsClientIdDeploymentsDeploymentIdStatus converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiV1ClientsClientIdDeploymentsDeploymentIdStatus(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "clientId" -------------
	var clientId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "clientId", runtime.ParamLocationPath, ctx.Param("clientId"), &clientId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter clientId: %s", err))
	}

	// ------------- Path parameter "deploymentId" -------------
	var deploymentId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "deploymentId", runtime.ParamLocationPath, ctx.Param("deploymentId"), &deploymentId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter deploymentId: %s", err))
	}

	ctx.Set(PayloadSignatureScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiV1ClientsClientIdDeploymentsDeploymentIdStatus(ctx, clientId, deploymentId)
	return err
}

// GetApiV1ClientsClientIdDeploymentsDeploymentIdDigest converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiV1ClientsClientIdDeploymentsDeploymentIdDigest(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "clientId" -------------
	var clientId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "clientId", runtime.ParamLocationPath, ctx.Param("clientId"), &clientId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter clientId: %s", err))
	}

	// ------------- Path parameter "deploymentId" -------------
	var deploymentId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "deploymentId", runtime.ParamLocationPath, ctx.Param("deploymentId"), &deploymentId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter deploymentId: %s", err))
	}

	// ------------- Path parameter "digest" -------------
	var digest string

	err = runtime.BindStyledParameterWithLocation("simple", false, "digest", runtime.ParamLocationPath, ctx.Param("digest"), &digest)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter digest: %s", err))
	}

	ctx.Set(PayloadSignatureScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiV1ClientsClientIdDeploymentsDeploymentIdDigestParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "If-None-Match" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("If-None-Match")]; found {
		var IfNoneMatch string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for If-None-Match, got %d", n))
		}

		err = runtime.BindStyledParameterWithLocation("simple", false, "If-None-Match", runtime.ParamLocationHeader, valueList[0], &IfNoneMatch)
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter If-None-Match: %s", err))
		}

		params.IfNoneMatch = &IfNoneMatch
	}
	// ------------- Optional header parameter "Accept-Encoding" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Accept-Encoding")]; found {
		var AcceptEncoding string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for Accept-Encoding, got %d", n))
		}

		err = runtime.BindStyledParameterWithLocation("simple", false, "Accept-Encoding", runtime.ParamLocationHeader, valueList[0], &AcceptEncoding)
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter Accept-Encoding: %s", err))
		}

		params.AcceptEncoding = &AcceptEncoding
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiV1ClientsClientIdDeploymentsDeploymentIdDigest(ctx, clientId, deploymentId, digest, params)
	return err
}

// PostApiV1Onboarding converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiV1Onboarding(ctx echo.Context) error {
	var err error

	ctx.Set(PayloadSignatureScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiV1Onboarding(ctx)
	return err
}

// GetApiV1OnboardingCertificate converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiV1OnboardingCertificate(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiV1OnboardingCertificate(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/api/v1/clients/:clientId/bundles/:digest", wrapper.GetApiV1ClientsClientIdBundlesDigest)
	router.POST(baseURL+"/api/v1/clients/:clientId/capabilities", wrapper.PostApiV1ClientsClientIdCapabilities)
	router.PUT(baseURL+"/api/v1/clients/:clientId/capabilities", wrapper.PutApiV1ClientsClientIdCapabilities)
	router.GET(baseURL+"/api/v1/clients/:clientId/deployments", wrapper.GetApiV1ClientsClientIdDeployments)
	router.POST(baseURL+"/api/v1/clients/:clientId/deployments/:deploymentId/status", wrapper.PostApiV1ClientsClientIdDeploymentsDeploymentIdStatus)
	router.GET(baseURL+"/api/v1/clients/:clientId/deployments/:deploymentId/:digest", wrapper.GetApiV1ClientsClientIdDeploymentsDeploymentIdDigest)
	router.POST(baseURL+"/api/v1/onboarding", wrapper.PostApiV1Onboarding)
	router.GET(baseURL+"/api/v1/onboarding/certificate", wrapper.GetApiV1OnboardingCertificate)

}